conversions: move more features inside the module
