+++ src/conversions/mod.rs
@@ -1,8 +1,9 @@
 //! This module contains conversions between various Rust object and their representation in Python.
 
 mod array;
-#[cfg(feature = "indexmap")]
-#[cfg_attr(docsrs, doc(cfg(feature = "indexmap")))]
+pub mod hashbrown;
 pub mod indexmap;
+pub mod num_bigint;
+pub mod num_complex;
 mod osstr;
 mod path;
