+++ src/conversions/num_bigint.rs
@@ -7,6 +7,7 @@
     docsrs,
     doc(cfg(all(feature = "num-bigint", not(any(Py_LIMITED_API, PyPy)))))
 )]
+
 //!  Conversions to and from [num-bigint](https://docs.rs/num-bigint)’s [`BigInt`] and [`BigUint`] types.
 //!
 //! This is useful for converting Python integers when they may not fit in Rust's built-in integer types.
